[Wallet] Fee fixes/refactor (#276)
### Description

**The problem**:
There is a single `suggestedFee` in the redux store used for both send and invite. 
Its value is not updated at appropriate times, or not at all, leading to issues like #36.
Also in `SendAmount`, the fee can be calculated too many times.

**This solution**:
The main idea is to have different types of (cached) estimated fees (right now for `send` and `invite`, later on for `exchange` and `escrow`) and trigger their update very infrequently depending on what flow the user is starting.
Then only when going to the `SendConfirmation` or `ReclaimPaymentConfirmationScreen` screens the actual fee is calculated once. Showing some feedback while it does it.

### Tested

Added/Updated tests.

Updated UI with loading/error feedback:
![Screenshot_1564563361](https://user-images.githubusercontent.com/57791/62292574-ffa24800-b466-11e9-987d-9cd82058b9a6.png)
![Screenshot_1564650868](https://user-images.githubusercontent.com/57791/62292684-3d06d580-b467-11e9-8469-746b5316bdeb.png)
![Screenshot_1564563456_censored(1)](https://user-images.githubusercontent.com/57791/62293130-36c52900-b468-11e9-9e6a-516f3985dae4.jpg)
![Screenshot_1564762306](https://user-images.githubusercontent.com/57791/62384908-e1664600-b553-11e9-84a9-14fd8f00f022.png)
![Screenshot_1564762309](https://user-images.githubusercontent.com/57791/62384916-e5926380-b553-11e9-8bfe-aa0285e11ed2.png)

### Other changes

N/A

### Related issues

- Fixes #36, #83, #113

### Backwards compatibility

Backward compatible.