whitelist oracle account and fund in the genesis block (#1725)
This does the setup for an oracle account on a network. It adds a new type of account that has its address and private key generated from the mnemonic. This account is included as an oracle for StableToken in the migration overrides that are set when deploying with celotool. By being included in that set, it is whitelisted as an oracle, and funded in the genesis block to cover transaction fees.