[dev-utils] Avoid devchain datadir corruption between test runs (#2776)
Copy the devchain for every test run. This way we avoid problems with tests that corrupt the chain.

Tested
Tested manually running the cli and contractKit tests

Other changes
Added the possibility to create the chain as a tar.gz. This allows in the future to keep in a storage different kinds of chains to be tested.

Related issues
Fixes #2761