Use KeyVault client for retrieving BLS key
## Description
The BLS key will be used to generate the salt for a given phone number mapping. As such, leaking this key can potentially compromise the privacy of users on the network. This change allows us to store the BLS key within a secure KeyVault instead of as an environment variable. 

## Other changes

Migrate the AzureKeyVaultClient to the contractkit utils directory since it's used by more than just account signing.

## Tested

Tested via mock object as well as using an actual KeyVault.

## Related issues

- Fixes https://github.com/celo-org/celo-labs/issues/429

## Backwards compatibility

These changes are not backwards compatible. We will need to deploy a KeyVault, create the corresponding key, and update the CloudFunction config before we can deploy the update.