[ReDesign] colors, fonts, Buttons
### Description

create versions of colors, fonts, and Button for the Mobile Redesign

### Other changes

While much of the existing interface remains, I removed the standard prop and added an optional size prop. 

I believe this is justified as it wasn't clear what standard meant. 

### Tested

I copied the existing test for Button, pointed it to v2 version and all the test passed. then because it was using `enzyme` i rewrote it to use `testing library`

### iOS 
![Screen Shot 2020-04-30 at 4 23 19 PM](https://user-images.githubusercontent.com/3814795/80769848-5e364880-8b03-11ea-90ac-548f7728b6fd.png)

### Android
![Screen Shot 2020-04-30 at 4 41 55 PM](https://user-images.githubusercontent.com/3814795/80769850-5f677580-8b03-11ea-8752-cc8236a51511.png)


### Related issues

- Fixes #3645 

### Backwards compatibility
no removed `standard` prop