(WIP) build: better selection of tests run in CI
### Description

![graph](https://user-images.githubusercontent.com/11168524/81671593-6ec2b900-9449-11ea-83fe-cabe209a90e3.jpg)

Here's a visualisation of our dependencies, this is now source controlled in the `dependency-graph.json` file. This PR improves our `ci_check_if_test_should_run_v2.sh` script so that we now only run tests for a package when:

- the yarn.lock file has changed
- the circleci config has changed
- the package has changed
- a dependency of the package has changed.

The `dependency-graph.json` file also gets automatically generated in CI and will exit early if it notices any differences from the checked in version. This is similar to how we cache our `node_modules` and confirm they match with our `yarn.lock`.

### Other changes

Moved some utility functions to their own file in the celotool repository. This is so we don't need to install the bls-zexe stuff just to run our `execCmd` helpers. 

### Tested

Ran a bunch of builds locally and on CI.

Working correctly in the following scenarios, with a dependency graph A->B->C (A depends on B depends on C).

- A has changed 
  - `./scripts/ci_check_if_test_should_run_v2.sh A` = true
  - `./scripts/ci_check_if_test_should_run_v2.sh B` = false
  - `./scripts/ci_check_if_test_should_run_v2.sh C` = false
- B has changed
  - `./scripts/ci_check_if_test_should_run_v2.sh A` = true
  - `./scripts/ci_check_if_test_should_run_v2.sh B` = true
  - `./scripts/ci_check_if_test_should_run_v2.sh C` = false
- C has changed
  - `./scripts/ci_check_if_test_should_run_v2.sh A` = true
  - `./scripts/ci_check_if_test_should_run_v2.sh B` = true
  - `./scripts/ci_check_if_test_should_run_v2.sh C` = true

### Related issues

- Fixes #3746