[Wallet] Adding recent payments to the state so we can check for 24 hour limit
## Description

This change adds a limit of $500 per 24 hours for making payments.
To accomplish this we need to store a list of recent payments with their timestamps, which we append to after payment is initiated. The interface then checks and displays an error if the limit is reached.

## Tested

Added unit tests for utils.

## Related issues

- fixes #3329 
- fixes https://github.com/celo-org/celo-monorepo/issues/3063

## Backwards compatibility

This change should be backwards compatible because we are only adding new properties to the state. We will not be able to check for payments sent before we started storing this so it will only function going forward. Since the limit is daily this should be fine.

## Commit Message

Adding recent payments to the redux state so we can check for 24 hour limit.