[contractKit][mobile] add percent of attestations completed as part of heuristic for phone number validation (will also be used in PGPNP) (#3829)
### Description

- Add percent of attestations completed as part of heuristic for phone number validation
- when new version of contractKit is published, this will also be used in PGPNP

`const DEFAULT_ATTESTATION_THRESHOLD = 0.8`
Value `.8` open for discussion, I feel this value is a bit aggressive.  I ran a script to calculate the average fraction of attestations completed from alfajores which resulted in `0.9153` with 479 identifiers.  Not sure how great of a proxy this is as many users are likely developers in areas where resources may be less constraint.

### Other changes

- Updated PGPNP get-salt to move the database call inside the try block in case of db failure

### Tested

Updated tests

### Related issues

- https://github.com/celo-org/celo-labs/issues/506

### Backwards compatibility

Backwards compatible change