[Wallet] Secure Send (#3791)
### Description
When there are multiple addresses associated with a single phone number, which can happen if someone gets SIM swapped or forgets their private key and opens a new account, a manual check by the user is needed to ensure they are transacting with the correct account. They do this by going through the Secure Send flow where they are asked to contact the recipient and confirm either the last 4 digits of their account number or the full thing. 

This comes into play in three different scenarios in which new addresses could be added to the phone number mapping:
1) Send payment
2) Request payment
3) Fulfill a payment request

### Other changes
This necessitated a rework of some of our data flow and logic so I refactored things as needed.

### Tested
Yes

### Related issues
- Fixes #1777 

### Backwards compatibility
Yes