[Wallet] Allow QR scan to circumvent Secure Send check (#3915)
### Description
Given QR scans are in person and have a lower likelihood of malicious activity, we want to minimize surface area for failures of QR scans so that it's as error-proof as possible. Quick fix to allow QR scans to circumvent the attestation contract lookup and thus bypass the SecureSend logic.

A more subtle change is that we are throwing an error if the recipient isn't doing a QR can and also doesn't have a phone number. In this case, we probably don't want the transaction to continue because we can't confirm the address mapping is up-to-date.

### Tested
Yep

### Related issues
Issue was raised in #3895 

### Backwards compatibility
Yes