Align unlock gold check to the smart contract implementation (#3914)
### Description

Smart contract and CLI implementations of the `unlock` gold requirements are currently inconsistent, as pointed out in https://github.com/celo-org/celo-monorepo/issues/3913. This PR aligns the check to the smart contract.

### Tested

Manual testing with/without the fix against a local testnet. Confirmed that the behavior described in #2861 and #3913 reproducible without the change and is fixed with the change.

### Related issues

- Fixes https://github.com/celo-org/celo-monorepo/issues/2861
- Fixes https://github.com/celo-org/celo-monorepo/issues/3913

### Backwards compatibility

100%