[Wallet] Fix withTranslation HOC typing (#4152)
### Description

This PR fixes the typing of our `withTranslation` HOC.

This is so we can still call hoisted static methods from the original component in a type safe way and not causing a TypeScript error.

This is useful in `Navigator` when we want to access `MyComp.navigationOptions`.

The small inconvenience of this is we have to explicitly pass the component `Props` when calling `withTranslation<Props>(...)`.
This is very similar to the `connect` HOC we already use.

I tried to lift this requirement but couldn't. Maybe if one day TypeScript supports generics of generics it could be inferred automatically. But for now I think it's good enough.

Note: this PR depends on #4145, please merge that one first.

### Tested

TypeScript passes successfully.

### Backwards compatibility

Yes.