[Wallet] Show errors during emulator installation (#4078)
### Description

After trying a billion combinations of parameters for running the emulator on circle ci we found something that runs without hiccups. It's still kind of slow (because of lack of hardware acceleration), but the plan is to try moving it to GitHub Actions which does provide some better hardware and the setup seems simpler overall.

### Tested

Tested locally and circle ci end to end test.

### Other changes

In the process I also updated detox, build tools (29), Android API (29), Xcode 11.5.0.

### Backwards compatibility

This change is backwards compatible.