[Wallet] Add additional security to wallet key storage (#4211)
### Description

To improve security of key storage, this generates a unique client pepper and store in the keystore/keychain to provide more entropy to the blockchain. It, combined with a user pin, is used to encrypt the private key on disk. This prevents an attacker with disk access from easily getting the private key

_[Rossy edit]:_

Also blacklisted unsafe pins, sanitized logs, fixed issue with pin caching while navigating, renamed lockWithPinEnabled to requirePinOnAppOpen, and some refactoring.

### Other changes

Store the mnemonic encrypted in the keystore.

### Tested

Run locally on android and iOS

_[Rossy edit]:_

Tested with adding, removing, and changing PIN settings on Android, confirmed the keystore values were still accessible. Mnemonic decrypts correctly and account can be unlocked.

Tested changing PIN on iOS, confirmed the same. Tested with no PIN on iOS with simulator, all good.