[Wallet] Custom brandings (#4282)
### Description

This implements custom brandings, along with a private branding for Valora.

Custom brandings are held in the `packages/mobiles/branding` folder. There are currently `celo` and `valora` brandings.
The `valora` branding is a private branding, held in a separate repository.

The idea is that each branding can override any of the files in `packages/mobile`.
A custom branding does that by defining the files it wants to override, following the same directory structure as the content of `packages/mobile`.

A postinstall script then rsync items into their final place based on the selected branding.

With this solution, files/assets that are branded need to be git ignored in their final location in the folder structure.
See `packages/mobile/.gitignore`.

By default the valora branding is selected when you run the sync script, it checkouts the private repo, and defaults to the celo branding if checkout fails.

You can manually switch branding:
```
# branding_name can be "valora" or "celo"
$ ./scripts/sync_branding -b branding_name
```

Make sure to select the `celo` branding before running tests to prevent changing snapshots.

The process for updating the private valora branding is as follow:
- make changes in `packages/mobile/branding/valora` and commit and push them to the private repo (master or PR, depending on the extent of the changes, this is very similar to maintaining a git submodule or nom package fork).
- update the `valora_branding_sha` in `packages/mobile/scripts/sync_branding.sh`
We can probably add tooling around this later on. Though these shouldn't change that much.

Note: celo icon is temporarily black because I haven't added an adaptive icon for it yet.

### Other changes

- Brandings can define branding specific config in `src/brandingConfig`, this is for the app name, email support etc...
- Removed hardcoded app name from i18n strings, `{{appName}}` is used instead.
- See valora specific changes (new icon, splash) in the private repo.
- Add `rsync` to `celohq:node10-gcloud:v3` docker image so CircleCI can successfully run `yarn install`

### Tested

Yes ;-)

### Related issues

- #3645 

### Backwards compatibility

Yes.