Create common PGPNP lib (#4278)
### Description

Creating a common library for the combiner and signer. This is important for having a shared interface that is kept in-sync across both services. Additionally, it allows for code-reuse and reduces duplication.

### Tested

Tested by integrating with combiner + signer (next PR).

### Backwards compatibility

The API object, SignMessageResponse, only adds fields:
  `performedQueryCount?: number`
  `totalQuota?: number`
and makes `signature` optional