[Wallet] Move signal handling to main (#4305)
### Description

Moving it earlier because I still see a few crashes coming in which seem to happen in the app delegate initialization code.

### Tested

It's difficult to trigger the actual crash that we see in production, which I suspect happens from one of the libraries that we use (one stack trace contained segment-firebase code). But I locally tested with the pkill command.

### Backwards compatibility

This change is backwards compatible.