Specify a context when deploying oracles, add 3rd mainnet cluster (#4299)
### Description

This does away with the "primary" and "secondary" terminology when referring to an oracle cluster, and instead introduces "contexts" for oracle deployments. An oracle context can really have any value, but in practice it should be: `<cloud provider>-<region>`, eg `azure-westus2` or `azure-westeurope`. This is what should be passed into celotool, for example:
```
celotooljs deploy initial oracle -e oracledev --context azure-eastus
```

Behind the scenes, this context is coerced into being all caps and using underscores instead of dashes. Env variables for a an oracle context are expected to be present and have a prefix of the form `ORACLE_<context in caps and underscores>_.*` (eg `ORACLE_AZURE_EASTUS_AZURE_SUBSCRIPTION_ID`). To do this, `DynamicEnvVar`s are introduced, which are a way of dynamically generating an environment variable name. For example:
```
ORACLE_AZURE_SUBSCRIPTION_ID = 'ORACLE_{{ oracleContext }}_AZURE_SUBSCRIPTION_ID',
``` 

In this case, `{{ oracleContext }}` is replaced at runtime with the value of whatever the oracleContext is.

This also moves oracle full node env vars to be context-specific, so that one context can have more/less full nodes or more/less storage per node.

### Other changes

* Adds a 3rd cluster in Azure that has full nodes constantly running as an emergency backup.
* Adds the ability to optionally specify which resource group a key vault is located in: `<address>:<key vault name>:<resource group (optional)>`

### Tested

Deployed!

### Related issues

n/a

### Backwards compatibility

Env variables must be updated