Signer DB Discrepancy Detection (#4334)
### Description

We'd like to be alerted when there is a discrepancy with the quota-related state in the databases of the signers. This change leverages the shared interface between the signer and combiner to pass the DB state information (performedQueryCount + totalQuota). 

### Other changes

Move shared constants and utils to the common library.

### Tested

Local unit tests.

### Backwards compatibility

Yes, the combiner shouldn't fail if the signer doesn't pass these new values.