[Wallet] Allow users to buy CELO on mainnet through Moonpay (#4508)
### Description

Update Moonpay widget URL to production, so that users can buy CELO in the Cash In flow. 

### Other changes

Added message explaining that Celo dollars are not yet available, but can be obtained by exchanging for CELO, based on a CUSD_MOONPAY_ENABLED feature flag:

![Screen Shot 2020-07-21 at 4 29 04 PM](https://user-images.githubusercontent.com/16907027/88117217-62588c00-cb6f-11ea-98c6-9133cae6c532.png)

This PR keeps the Cash In input as the local currency displayed in dollars- still in discussion with @nityas on a possible change here

Added signMoonpayStaging to maintain the old staging API keys and setup that we will want to use to test the cUSD integration. 

### Tested

Confirmed that Moonpay widget loaded with expected values pre-filled with the new `signMoonpayProd` function deployed to `celo-org-mainnet`. Confirmed that API keys from the encrypted .env file were properly loaded to firebase. 

Cannot test full Moonpay flow as CELO is not available in the US via Moonpay (waiting to hear back from them on what is holding this up). 

### Related issues

- Fixes #4214 

### Backwards compatibility

Yes, the previous signMoonpay function remains deployed.