Add transaction confirmed event (#4568)
### Description

Currently, `transaction_start` is emitted when a transaction confirmation is returned from `web3`.
This causes two `transaction_start` events to appear in the analytics database for each transaction, in addition to not being the intended semantics event.

This PR adds a `transaction_confirmed` event which is emitted when a transaction confirmation is received from `web3`