Add transferAttestationMapping to Attestations.sol (#4583)
### Description

As part of the initiative for [Feeless Attestations with Meta Transactions](https://docs.google.com/document/d/1wBFX6c1Tu4Dyxrx9JdKhmXPXXaFmVpHmP5bAAzmX6EU/edit#heading=h.79symh8gkwep), we need to be able to migrate attestations between Valora externally owned account and the smart wallet (or vice versa if Valora chooses not to be backed by the smart wallet and we only use the smart wallet as a mean for covering attestations fees)

NOTE: I followed the `revoke` function's pattern of requiring the index of the address as a parameter.  I'm assuming this is a pattern to reduce transaction cost as the caller can make a view call to `lookupAccountsForIdentifier` and determine the index themselves.

### Other changes

NA

### Tested

Added Test

### Related issues

[Feeless Attestations with Meta Transactions](https://docs.google.com/document/d/1wBFX6c1Tu4Dyxrx9JdKhmXPXXaFmVpHmP5bAAzmX6EU/edit#heading=h.79symh8gkwep)

### Backwards compatibility

Requires smart contract release -- adds new external function.  Does not impact current storage or existing functions.