Add network events to mobile Analytics package (#4599)
### Description

In order to get a better picture for how long users are waiting to become synced to the network, this PR added a number of events around the connection and sync status of the mobile app to the Celo network.

New events included here are:

  * `network_connected`
  * `network_disconnected`
  * `network_sync_waiting`
  * `network_sync_start`
  * `network_sync_finish`
  * `network_sync_error`
  * `network_sync_lost`
  * `network_sync_restored`

Additionally, this PR changes syncing logic to consider any block older than 60 seconds to be stale, and wait to catch up. This is important because it is possible to be behind the head of the chain, but not syncing. This matches the logic used in the `celocli` https://github.com/celo-org/celo-monorepo/blob/6e9c7dbb985947eb2d680b3a9e18de5acc22e333/packages/cli/src/utils/helpers.ts#L21

It is important to catch this edge case because if a transaction is submitted while not syncing, it will not get a receipt and will wait 5 minutes then report the transaction failed. Or if the client starts to sync while waiting for a receipt, and the chain advances more than 50 blocks, the error `Transaction was not mined within 50 blocks` will be reported. It's possible this is related to https://github.com/celo-org/celo-blockchain/issues/1111, but the only evidence here is that _if_ this is happening, the error would be the same.

In order to track sync status, known chain head information is added to the redux state. This information could potentially be leveraged for other uses, but at the moment is only used to report events to the analytics platform.

### Other changes

* Add a `create_geth_error` event to track when the app fails to create a Geth instance.
* Add extra data fields to various transaction events to enable better tracking

### Tested

Unit test for the syncing logic change. Otherwise untested.

### Backwards compatibility

100%