Prevent nonce assignment collisions when sending multiple transactions (#4678)
### Description

When sending multiple transactions at once, if signing a transaction and adding it to the transaction pool are not done atomically, then multiple of the transactions may be assigned the same nonce. Currently we see this issue when trying to complete the verification flow, where 3 transactions are submitted in parallel, one for each attestation to be completed. https://github.com/celo-org/celo-blockchain/issues/1111

This PR adds a simple implementation of a mutex lock, and uses it to ensure that once a nonce is assigned to a transaction, no other transaction can have a nonce assigned until that transaction is in the node transaction pool, or has been rejected by the node.

### Other changes

* Changed initialization of Geth to allow for fetching static nodes file and genesis file in parallel.

### Tested

Added a unit test for the lock, and used a manual test recreating the error observed during verification completion to ensure that error was fixed.

### Related issues

- Fixes https://github.com/celo-org/celo-blockchain/issues/1111

### Backwards compatibility

100%