[Wallet] Implement remove account feature (#4509)
### Description

Added a button to allow removing account from the device. The user is asked for their Account Key before being allowed to do it.
Some screenshots of the flow:
<img width="348" alt="Screen Shot 2020-07-22 at 1 25 48 AM" src="https://user-images.githubusercontent.com/6062888/88134300-a329c280-cbba-11ea-8b23-f5ebfdfce38d.png">
<img width="325" alt="Screen Shot 2020-07-22 at 1 25 54 AM" src="https://user-images.githubusercontent.com/6062888/88134311-a624b300-cbba-11ea-8749-1fe79e0b0c72.png">
<img width="345" alt="Screen Shot 2020-07-22 at 1 26 20 AM" src="https://user-images.githubusercontent.com/6062888/88134315-a755e000-cbba-11ea-85cf-52b451d72a2e.png">
<img width="329" alt="Screen Shot 2020-07-22 at 1 27 26 AM" src="https://user-images.githubusercontent.com/6062888/88134318-a7ee7680-cbba-11ea-9c97-386c0a553748.png">
<img width="334" alt="Screen Shot 2020-07-22 at 1 27 35 AM" src="https://user-images.githubusercontent.com/6062888/88134319-a7ee7680-cbba-11ea-9ada-e265ef2b638b.png">


### Other changes

There were some dialogs that were using a `WarningModal` which had an old design. Replaced it with `Dialog` components that have the latest UI changes.

### Tested

Tested manually to remove the account and created a new one to verify that my account data was deleted.