Ensure every transaction has a unique ID for logging (#4711)
### Description

Currently in the mobile application transactions are IDs when created, but in many cases these IDs are not unique or descriptive enough to provide context, or track an individual transaction.
In trying to track the performance of transaction, and diagnose errors from logging and analytics, this is a challenge.

This PR adds a new `TransactionContext` interface and ensures that every transaction has a unique ID and short description.

### Tested

Updated unit tests and manual testing of the application on my device.

### Backwards compatibility

Transaction ID format changed from a SHA256 hash to a UUID.