[Wallet] Payment requests must use lowercase addresses (#4973)
### Description

When receiving a payment request, we look on Firebase for requests made to our address using lowercase letters. If the request was made using uppercase letters, we ignore that request. With this change, we are making sure requests are always made using lowercase addresses.

### Tested

Sent a payment request without this change, it didn't arrive. Made the change, made a new request, it arrived. 
Also did the same thing but manually changing from upper to lowercase on Firebase and the requests started showing up after doing so.

### Related issues

- Fixes #4886

### Backwards compatibility

Requests that failed in the past won't start working, new payment requests will need to be made. I think that's for the better though.