[Wallet] Fix bug to allow users to skip invite redemption after canceling (#4977)
### Description
If a user presses "Cancel" while an invite redemption is in progress, it should reset the invite redemption state variables. Currently, `isRedemeeingInvite` is never flipped back to `false`. Additionally, when a invite redemption fails it often does so after an account has been created but before the funds have been transferred from the temp invite account to the user's account. The combination of `account` and `isRedeemingInvite` being truthy results in the user being stuck on the `EnterInviteCode` screen indefinitely.

### Other changes
- When a user copy/pastes the full invite text message, the input form used to get populated with the private key that was extracted from the parsed invite code. This was confusing users so instead we are now populating the input form with the parsed invite code instead of the private key
- Added a cancel action, which gets triggered when the user presses "cancel" on the `EnterInviteCode` screen,  to the send transaction within `moveAllFundsFromAccount` 
- Changed the condition by which `addTempAccountToWallet` detects whether an account has already been created. I think this may have gotten overlooked during the recent geth changes

### Tested
Added tests and ran through various conditions on a physical device.

### Related issues
NA

### Backwards compatibility
Yes