[Wallet] Update Clipboard for iOS 14 (#5121)
### Description

On iOS 14 when we read the clipboard the user gets a notification. Since we're reading it every second to detect changes, it's quite annoying. This PR's goal is to have everything work the way it used to except on iOS devices running version 14+. In those devices, the paste button/icons are always visible if the clipboard is not empty and is hidden only if the clipboard is empty. In short, we lose the "smart" paste icons on iOS 14+ that showed up only if the format of the content of the clipboard matches the expected format.

### Other changes

Updated the clipboard library.

### Tested

On the simulator using different iOS versions.

### Related issues

- Fixes #5120

Part of this epic: https://github.com/celo-org/celo-monorepo/issues/4835