Verify source code matches deployed or proposed bytecode (#4996)
### Description

This PR adds a script that verifies that a given set of Solidity sources matches bytecodes deployed to a Celo network. An optional `proposal.json` can also be passed in, representing proposed changes that are not yet reflected by onchain Registry/proxies' state.

This check will rely on:
* Each (non-library) core contract being registered in the Registry
* Each linked library being proxied

The deployed Celo contracts do not currently satisfy these assumptions, so there is some temporary code included to handle the current case. The checks are also not 100% complete for the current RC1 deployment (e.g. the Registry itself is missed since it's not registered).

The main check happens in the `verifyBytecodes` function, which takes a list of core contract names and a build directory as input. The DFS works as follows:

1. If we're on a core contract, look up its address in the Registry. We keep track of library addresses as we encounter them.
2. Find all instances of `__LibraryName__...` in the compiled bytecode.
3. Ensure that each `__LibraryName__...` link corresponds to the same address in the onchain bytecode.
4. Link the compiled bytecode and verify it matches the onchain bytecode.
5. Add any freshly encountered libraries to the DFS stack.

The biggest remaining piece of work on this front is also verifying *proposed* changes, that is, given a Celo network and a Governance proposal, verifying that the given sources correspond to what the Celo system would look like after applying the proposal.

### Tested

Unit tests for the library doing the heavy lifting.

For a manual test against baklava,

    git checkout baklava
    yarn build
    mv build/contracts build/baklava-contracts
    git checkout m-chrzan/deployed-bytecode-check
    # point the script to build/baklava-contracts
    yarn build:ts && yarn truffle exec --network baklava scripts/truffle/verify-bytecode.js

### Other changes

Moved some duplicated code out into a new `lib/bytecode` library.

### Related issues

- Fixes #4496 
- Fixes #4817

### Backwards compatibility

Very much so.