Select static node based on region resolved from timezone (#5266)
### Description

Building on the recent work to deploy Forno full nodes to multiple regions (e.g. https://github.com/celo-org/celo-monorepo/pull/5113), this PR adds a support in contractkit for simple resolution of the users timezone to inform what region they are in. (Note that this method is by no means perfect, but it should work in most cases) On mainnet, and other networks if more regions are deployed in the future, a region-specific `static-node.json` file is fetched from GCP to have the app establish connection to relatively nearby set of peers.

### Other changes

* Add the latest block number to the `network_sync_waiting` Segment event

### Tested

Added a number of new unit tests. Manually ran the new code on a development build and confirmed it connected to the right nodes for my region. Also changed by timezone and observed that it connected to the correct nodes for the region of the new timezone.