Make initialization of the proxy part of the governance proposal (#5481)
### Description

This PR removes the initialization of the proxy in `make-release` and instead makes it part of the proposal. This allows the initialization be more verifiable instead of relying on the deployer of the proxy to have initialized it correctly.

For that, we have to skip the setting of the implementation on the proxy, as otherwise any caller to `initialize` would be able to make themselves owner. That also means that we are no longer transferring ownership with respect to the implementation contract functions in proxy storage.

A follow up PR should do the [verification of the initialization transaction](https://github.com/celo-org/celo-monorepo/issues/5484) as well as [verifying that the storage root is "pure"](https://github.com/celo-org/celo-monorepo/issues/5483)

This PR also fixes a grave issue where we are passing the hashed contract name to `Registry#setAddressFor` when in reality, it takes the plain contract name and hashes on-chain.

Fixes #5479