Update Governance documentation to reflect Mainnet (#5512)
### Description

While reviewing the [Celo Community Fund Constitution proposal](https://docs.google.com/document/d/15xkmj6mXaLAjNcdvGX3PKGM3aanuvVL1IYXeWHvfxBs/edit) I noiced a number of errors in that document that were sourced from our own Governance documentation.

In order to correct these errors, and avoid future confusion, this PR includes a number of edits to the Governance documentation to bring it inline with what is deployed to Mainnet and to increase the clarity to the reader.