[cli] Require --force with account:claim-attestation-service-url for non-TLS urls (#5599)
### Description

In the future we will be requiring TLS for Attestation Service URLs. This adds a check to the CLI command to require `--force` to add a URL that doesn't begin `https://`.
 
### Tested

Locally

### Backwards compatibility

In the unlikely event that users have scripted this, and do not have TLS deployed, scripts will need modification.